Finalise the command queue system.
Command queue system finalised and working.
Removed frame delay on nearest target acquisition.
