+++ Assets/Scripts/Components/Authoring/UnitAuthoringComponent.cs
@@ -26,7 +26,7 @@ public class UnitAuthoringComponent : MonoBehaviour, IConvertGameObjectToEntity
 		dstManager.AddComponentData(entity, freezeRotation);
 		dstManager.AddComponentData(entity, new UnitTag());
 
-		dstManager.AddComponentData(entity, new CurrentTarget{ findTargetOfType = AITargetType.None, targetData = new TargetData()});
+		dstManager.AddComponentData(entity, new CurrentTarget{ targetData = new TargetData()});
 		dstManager.AddComponentData(entity, new PreviousTarget{ targetData = new TargetData()});
 
 		if((unitType & UnitType.Harvester) != 0)
