+++ Assets/Scripts/Components/Command.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ad06770e70111b1e8bfe2ae7d4c7403b
+guid: e7a333ee0971f154f9add45d175a38e2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
